Added working update_thumbs.php feature and moved the createthumb code into a separate function.

+++ update_thumbs.php
@@ -1,20 +1,30 @@
 <?php
-	require("config_default.php");
-	include("config.php");
+	require('config_default.php');
+	include('config.php');
+	require('functions.php');
 
-	$thumbs_dir = 'thumbs';
-	
-	// Empty /thumbs folder
-	foreach (scandir($thumbs_dir) as $item) {
-		if ($item == '.' || $item == '..') continue;
-		unlink($thumbs_dir.DIRECTORY_SEPARATOR.$item);
+	$path_to_scan = "photos/";
+	if (isset($_GET['path'])) {
+		$path_to_scan .= $_GET['path']; // TODO: remove '..'
 	}
-	
+	else {
+		// If global scan, we can empty the thumbs folder
+		foreach (scandir($thumbs_dir) as $item) {
+			if ($item == '.' || $item == '..') continue;
+			unlink($thumbs_dir.DIRECTORY_SEPARATOR.$item);
+		}
+	}
+
 	// Build thumbs
-	$file_names = scanFileNameRecursivly('photos');
+	$file_names = scanFileNameRecursivly($path_to_scan);
 	foreach ($file_names as $file_name) {
-		fopen();
-		//echo '<img src="createthumb.php?filename='.$file_name.'&amp;size='.$thumb_size.'">';
+		if ($file_name[strlen($file_name)-1] != '.') {
+			//echo $file_name.'<br>';
+			createthumb($file_name, $thumb_size, false);
+			echo '.';
+			flush();
+			ob_flush();
+		}
 	}
 	
 function scanFileNameRecursivly($path = '', &$name = array() )
